refactor: move DEFAULT_TIME_FORMAT to src/utils/datetime/constants
