+++ src/functions/components/FunctionRunsRow.tsx
@@ -13,7 +13,7 @@ import FunctionRunLogsOverlay from 'src/functions/components/FunctionRunLogsOver
 
 // Types
 import {FunctionRun} from 'src/client/managedFunctionsRoutes'
-import {DEFAULT_TIME_FORMAT} from 'src/shared/constants'
+import {DEFAULT_TIME_FORMAT} from 'src/utils/datetime/constants'
 
 // Utils
 import {FormattedDateTime} from 'src/utils/datetime/FormattedDateTime'
