+++ src/shared/constants/index.ts
@@ -41,8 +41,6 @@ export const EMPTY_ZUORA_PARAMS: CreditCardParams = {
 export const BALANCE_THRESHOLD_DEFAULT = 10
 export const MINIMUM_BALANCE_THRESHOLD = 1
 
-export const DEFAULT_TIME_FORMAT = 'YYYY-MM-DD HH:mm:ss'
-
 export const DROPDOWN_MENU_MAX_HEIGHT = 240
 
 export const PRESENTATION_MODE_ANIMATION_DELAY = 0 // In milliseconds.
