+++ src/utils/datetime/README.md
@@ -6,7 +6,7 @@ Our UI's Canonical DateTime formatter. It comes with a component which is aware 
 
 ```ts
 import {FormattedDateTime} from 'src/utils/datetime/FormattedDateTime'
-import {DEFAULT_TIME_FORMAT} from 'src/shared/constants/index'
+import {DEFAULT_TIME_FORMAT} from 'src/utils/datetime/constants'
 
 const FunctionFella: ({timestamp: string}) => {
   return <FormattedDateTime format={DEFAULT_TIME_FORMAT} date={new Date(timestamp)} />
