+++ src/views/selectors/index.ts
@@ -3,7 +3,7 @@ import {AppState, View, ResourceType, Dashboard} from 'src/types'
 
 // Selectors
 import {getByID} from 'src/resources/selectors'
-import {DEFAULT_TIME_FORMAT} from 'src/shared/constants'
+import {DEFAULT_TIME_FORMAT} from 'src/utils/datetime/constants'
 
 export const getViewsForDashboard = (
   state: AppState,
