+++ src/visualization/types/SingleStat/view.tsx
@@ -22,7 +22,7 @@ import {
 
 import './style.scss'
 import {AppSettingContext} from 'src/shared/contexts/app'
-import {DEFAULT_TIME_FORMAT} from 'src/shared/constants'
+import {DEFAULT_TIME_FORMAT} from 'src/utils/datetime/constants'
 
 // Selectors
 import {isAnnotationsModeEnabled} from 'src/annotations/selectors'
